requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.6.0
beanie==1.23.6
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic-settings==2.1.0
pydantic[email]==2.5.2
websockets==12.0
